+++ extensions/audio-device/src/helpers.tsx
@@ -1,6 +1,5 @@
 import {
   ActionPanel,
-  Clipboard,
   closeMainWindow,
   Color,
   getPreferenceValues,
@@ -11,8 +10,9 @@ import {
   showToast,
   Toast,
   Action,
+  Keyboard,
 } from "@raycast/api";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import {
   AudioDevice,
   getInputDevices,
@@ -24,42 +24,8 @@ import {
   setDefaultSystemDevice,
   TransportType,
 } from "./audio-device";
-import { createDeepLink, setOutputDevice } from "./utils";
-
-type UseAudioDevicesResponse = {
-  isLoading: boolean;
-  data: {
-    devices: AudioDevice[];
-    current: AudioDevice;
-  };
-};
-
-export function useAudioDevices(type: "input" | "output") {
-  const [isLoading, setIsLoading] = useState(true);
-  const [audioDevices, setAudioDevices] = useState<UseAudioDevicesResponse["data"]>();
-
-  useEffect(() => {
-    const fetchDevices = async () => {
-      const devices = await (type === "input" ? getInputDevices() : getOutputDevices());
-      const current = await (type === "input" ? getDefaultInputDevice() : getDefaultOutputDevice());
-
-      return {
-        devices,
-        current,
-      };
-    };
-
-    fetchDevices()
-      .then(setAudioDevices)
-      .catch((err) => showToast(Toast.Style.Failure, `There was an error fetching the audio devices.`, err.message))
-      .finally(() => setIsLoading(false));
-  }, [type]);
-
-  return {
-    isLoading,
-    data: audioDevices,
-  };
-}
+import { createDeepLink } from "./utils";
+import { usePromise } from "@raycast/utils";
 
 type DeviceListProps = {
   type: "input" | "output";
@@ -68,8 +34,6 @@ type DeviceListProps = {
 
 export function DeviceList({ type, deviceId }: DeviceListProps) {
   const { isLoading, data } = useAudioDevices(type);
-  const subtitle = (device: AudioDevice) =>
-    Object.entries(TransportType).find(([, v]) => v === device.transportType)?.[0];
 
   useEffect(() => {
     if (!deviceId || !data?.devices) return;
@@ -90,7 +54,7 @@ export function DeviceList({ type, deviceId }: DeviceListProps) {
         showToast(
           Toast.Style.Failure,
           `Error!`,
-          `There was an error setting the active ${type} audio device to ${device.name}`
+          `There was an error setting the active ${type} audio device to ${device.name}`,
         );
       }
     })();
@@ -99,44 +63,52 @@ export function DeviceList({ type, deviceId }: DeviceListProps) {
   return (
     <List isLoading={isLoading}>
       {data &&
-        data.devices.map((d) => (
-          <List.Item
-            key={d.uid}
-            title={d.name}
-            subtitle={subtitle(d)}
-            icon={deviceIcon(d)}
-            actions={
-              <ActionPanel>
-                <SetAudioDeviceAction device={d} type={type} />
-                <Action.CreateQuicklink
-                  title="Create quicklink"
-                  quicklink={{
-                    name: `Set ${d.isOutput ? "Output" : "Input"} Device to ${d.name}`,
-                    link: createDeepLink(d.isOutput ? "set-output-device" : "set-input-device", {
-                      deviceId: d.id,
-                    }),
-                  }}
-                />
-                <Action
-                  title={`Copy Device Name to Clipboard`}
-                  onAction={async () => {
-                    await Clipboard.copy(d.name);
-                    await showToast({
-                      style: Toast.Style.Success,
-                      title: "Device name copied to the clipboard",
-                    });
-                  }}
-                />
-              </ActionPanel>
-            }
-            accessories={[
-              {
-                icon: d.uid === data.current.uid ? Icon.Checkmark : undefined,
-              },
-            ]}
-          />
-        ))}
+        data.devices.map((d) => {
+          const isCurrent = d.uid === data.current.uid;
+          return (
+            <List.Item
+              key={d.uid}
+              title={d.name}
+              subtitle={getSubtitle(d)}
+              icon={getIcon(d, d.uid === data.current.uid)}
+              actions={
+                <ActionPanel>
+                  <SetAudioDeviceAction device={d} type={type} />
+                  <Action.CreateQuicklink
+                    quicklink={{
+                      name: `Set ${d.isOutput ? "Output" : "Input"} Device to ${d.name}`,
+                      link: createDeepLink(d.isOutput ? "set-output-device" : "set-input-device", {
+                        deviceId: d.id,
+                      }),
+                    }}
+                  />
+                  <Action.CopyToClipboard
+                    title="Copy Device Name"
+                    content={d.name}
+                    shortcut={Keyboard.Shortcut.Common.Copy}
+                  />
+                </ActionPanel>
+              }
+              accessories={getAccessories(isCurrent)}
+            />
+          );
+        })}
     </List>
+  );
+}
+
+function useAudioDevices(type: "input" | "output") {
+  return usePromise(
+    async (type) => {
+      const devices = await (type === "input" ? getInputDevices() : getOutputDevices());
+      const current = await (type === "input" ? getDefaultInputDevice() : getDefaultOutputDevice());
+
+      return {
+        devices,
+        current,
+      };
+    },
+    [type],
   );
 }
 
@@ -148,20 +120,17 @@ type SetAudioDeviceActionProps = {
 function SetAudioDeviceAction({ device, type }: SetAudioDeviceActionProps) {
   return (
     <Action
-      title={`Select ${device.name}`}
+      title={`Set as ${type === "input" ? "Input" : "Output"} Device`}
+      icon={{ source: type === "input" ? "mic.png" : "speaker.png", tintColor: Color.PrimaryText }}
       onAction={async () => {
         try {
           await (type === "input" ? setDefaultInputDevice(device.id) : setOutputAndSystemDevice(device.id));
           closeMainWindow({ clearRootSearch: true });
           popToRoot({ clearSearchBar: true });
-          showHUD(`Active ${type} audio device set to ${device.name}`);
+          showHUD(`Set "${device.name}" as ${type} device`);
         } catch (e) {
           console.log(e);
-          showToast(
-            Toast.Style.Failure,
-            `Error!`,
-            `There was an error setting the active ${type} audio device to ${device.name}`
-          );
+          showToast(Toast.Style.Failure, `Failed setting "${device.name}" as ${type} device`);
         }
       }}
     />
@@ -176,9 +145,21 @@ async function setOutputAndSystemDevice(deviceId: string) {
   }
 }
 
-export function deviceIcon(device: AudioDevice) {
+function getIcon(device: AudioDevice, isCurrent: boolean) {
   return {
     source: device.isInput ? "mic.png" : "speaker.png",
-    tintColor: Color.SecondaryText,
+    tintColor: isCurrent ? Color.Green : Color.SecondaryText,
   };
+}
+
+function getAccessories(isCurrent: boolean) {
+  return [
+    {
+      icon: isCurrent ? Icon.Checkmark : undefined,
+    },
+  ];
+}
+
+function getSubtitle(device: AudioDevice) {
+  return Object.entries(TransportType).find(([, v]) => v === device.transportType)?.[0];
 }