+++ extensions/doutu/src/services/clipboardService.ts
@@ -2,6 +2,7 @@ import { environment, showHUD, showToast, Toast } from "@raycast/api";
 import fetch from "node-fetch";
 import fs from "fs";
 import proc from "child_process";
+import imageType from "image-type";
 
 export default {
   imageToClipboard: async (imageUrl: string) => {
@@ -21,9 +22,16 @@ export default {
     const fileName = imageUrl.substring(imageUrl.lastIndexOf("/") + 1);
     const response = await fetch(imageUrl);
     const data = await response.arrayBuffer();
-    const filePath = `${tempDir}/${fileName}`;
-    await fs.writeFileSync(filePath, new Uint8Array(data), "binary");
-    proc.exec(`osascript -e 'set the clipboard to (read (POSIX file "${filePath}") as JPEG picture)'`);
+    const buffer = new Uint8Array(data);
+    const type = await imageType(buffer);
+    let file = `${tempDir}/${fileName}`;
+    if (!fileName.endsWith(`.${type?.ext}`)) file += `.${type?.ext}`;
+    await fs.writeFileSync(file, new Uint8Array(data), "binary");
+    if (type?.mime === "image/gif") {
+      proc.exec(`osascript -e 'set the clipboard to POSIX file "${file}"'`);
+    } else {
+      proc.exec(`osascript -e 'set the clipboard to (read (POSIX file "${file}") as JPEG picture)'`);
+    }
     toast.hide();
     showHUD("✅ Copied to clipboard");
   },