+++ extensions/github/src/unread-notifications.tsx
@@ -11,8 +11,8 @@ import {
 } from "@raycast/api";
 import { useCachedPromise } from "@raycast/utils";
 
+import { getGitHubClient } from "./api/githubClient";
 import { Notification } from "./components/NotificationListItem";
-import View from "./components/View";
 import {
   getGitHubIcon,
   getGitHubURL,
@@ -20,7 +20,7 @@ import {
   getNotificationSubtitle,
   getNotificationTooltip,
 } from "./helpers/notifications";
-import { getGitHubClient } from "./helpers/withGithubClient";
+import { withGitHubClient } from "./helpers/withGithubClient";
 import { useViewer } from "./hooks/useViewer";
 
 const preferences = getPreferenceValues<Preferences.UnreadNotifications>();
@@ -162,10 +162,4 @@ function UnreadNotifications() {
   );
 }
 
-export default function Command() {
-  return (
-    <View>
-      <UnreadNotifications />
-    </View>
-  );
-}
+export default withGitHubClient(UnreadNotifications);